resolve feedback: move out file, add a11y role, remove unnecesary code
