--- src/components/app-bar/__tests__/app-bar.atoms.test.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react'
-import { NavSearchButton } from '../app-bar.atoms'
-
-describe('NavSearchButton', () => {
-  it('should match snapshot', () => {
-    const { asFragment } = render(<NavSearchButton />)
-
-    expect(asFragment()).toMatchSnapshot()
-  })
-})
