+++ src/components/nav-search-button/__tests__/nav-search-button.test.tsx
@@ -1,0 +1,16 @@
+import { render } from '@testing-library/react'
+import { NavSearchButton } from '../nav-search-button'
+
+describe('NavSearchButton', () => {
+  it('should match snapshot', () => {
+    const { asFragment } = render(<NavSearchButton />)
+
+    expect(asFragment()).toMatchSnapshot()
+  })
+
+  it('render without shortcut and match snapshot', () => {
+    const { asFragment } = render(<NavSearchButton useShortcut={false} />)
+
+    expect(asFragment()).toMatchSnapshot()
+  })
+})
